Moved Vec3d struct into utils
