+++ include/States.h
@@ -4,29 +4,11 @@
 #include<vector>
 #include <stdexcept>
 #include <gsl/gsl_interp.h>
-#include <ale.h> 
+#include <ale.h>
+
+#include "Util.h"
 
 namespace ale {
-
-/** A 3D cartesian vector */
-struct Vec3d {
-  double x;
-  double y;
-  double z;
-
-  // Accepts an {x,y,z} vector
-  Vec3d(const std::vector<double>& vec) {
-    if (vec.size() != 3) {
-      throw std::invalid_argument("Input vector must have 3 entries.");
-    }
-    x = vec[0];
-    y = vec[1];
-    z = vec[2];
-  }; 
-
-  Vec3d(double x, double y, double z) : x(x), y(y), z(z) {};
-  Vec3d() : x(0.0), y(0.0), z(0.0) {};
-};
 
 /** A state vector with position and velocity*/
 struct State {
@@ -42,7 +24,7 @@ struct State {
     velocity = {vec[3], vec[4], vec[5]};
   };
 
-  State(ale::Vec3d position, ale::Vec3d velocity) : position(position), velocity(velocity) {}; 
+  State(ale::Vec3d position, ale::Vec3d velocity) : position(position), velocity(velocity) {};
   State() {};
 };
 
@@ -50,19 +32,19 @@ class States {
   public:
     // Constructors
     States();
-    States(const std::vector<double>& ephemTimes, const std::vector<ale::Vec3d>& positions, 
+    States(const std::vector<double>& ephemTimes, const std::vector<ale::Vec3d>& positions,
            int refFrame=1);
 
-    States(const std::vector<double>& ephemTimes, const std::vector<ale::Vec3d>& positions, 
+    States(const std::vector<double>& ephemTimes, const std::vector<ale::Vec3d>& positions,
            const std::vector<ale::Vec3d>& velocities, int refFrame=1);
 
-    States(const std::vector<double>& ephemTimes, const std::vector<ale::State>& states, 
-           int refFrame=1); 
+    States(const std::vector<double>& ephemTimes, const std::vector<ale::State>& states,
+           int refFrame=1);
 
     ~States();
-    
+
     // Getters
-    std::vector<ale::State> getStates() const; //! Returns state vectors (6-element positions&velocities) 
+    std::vector<ale::State> getStates() const; //! Returns state vectors (6-element positions&velocities)
     std::vector<ale::Vec3d> getPositions() const; //! Returns the current positions
     std::vector<ale::Vec3d> getVelocities() const; //! Returns the current velocities
     std::vector<double> getTimes() const; //! Returns the current times
@@ -71,14 +53,14 @@ class States {
 
   /**
    * Returns a single state by interpolating state.
-   * If the Cache has been minimized, a cubic hermite is used to interpolate the 
-   * position and velocity over the reduced cache. 
+   * If the Cache has been minimized, a cubic hermite is used to interpolate the
+   * position and velocity over the reduced cache.
    * If not, a standard gsl interpolation will be done.
-   * 
+   *
    * @param time Time to get a value at
    * @param interp Interpolation type to use. Will be ignored if cache is minimized.
-   * 
-   * @return ale::State 
+   *
+   * @return ale::State
    */
     ale::State getState(double time, ale::interpolation interp=LINEAR) const;
 
@@ -89,45 +71,44 @@ class States {
     ale::Vec3d getVelocity(double time, ale::interpolation interp=LINEAR) const;
 
     /** Returns the first ephemeris time **/
-    double getStartTime(); 
+    double getStartTime();
 
     /** Returns the last ephemeris time **/
-    double getStopTime(); 
-    
+    double getStopTime();
+
   /**
    * Perform a cache reduction. After running this, getStates(), getPositions(),
    * and getVelocities() will return vectors of reduced size, and getState(),
    * getPosition(), and getVelocity() will
-   * returns values interpolated over the reduced cache using a cubic hermite spline 
-   *  
-   * Adapted from Isis::SpicePosition::reduceCache().  
-   * 
+   * returns values interpolated over the reduced cache using a cubic hermite spline
+   *
+   * Adapted from Isis::SpicePosition::reduceCache().
+   *
    * @param tolerance Maximum error between hermite approximation and original value.
    */
     States minimizeCache(double tolerance=0.01);
 
   private:
-  
+
   /**
    * Calculates the points (indicies) which need to be kept for the hermite spline to
-   * interpolate between to mantain a maximum error of tolerance. 
-   *  
-   * Adapted from Isis::SpicePosition::HermiteIndices. 
-   *  
+   * interpolate between to mantain a maximum error of tolerance.
+   *
+   * Adapted from Isis::SpicePosition::HermiteIndices.
+   *
    * @param tolerance Maximum error between hermite approximation and original value.
    * @param indexList The list of indicies that need to be kept.
    * @param baseTime Scaled base time for fit
    * @param timeScale Time scale for fit.
-   * 
-   * @return std::vector<int> 
+   *
+   * @return std::vector<int>
    */
-    std::vector<int> hermiteIndices(double tolerance, std::vector <int> indexList, 
+    std::vector<int> hermiteIndices(double tolerance, std::vector <int> indexList,
                                     double baseTime, double timeScale);
     std::vector<ale::State> m_states; //! Represent at states internally to keep pos, vel together
     std::vector<double> m_ephemTimes; //! Time in seconds
-    int m_refFrame;  //! Naif IDs for reference frames 
+    int m_refFrame;  //! Naif IDs for reference frames
   };
 }
 
 #endif
-
